ref(inspection): share the body-symbol walk between the unused finders
The let-binding finder and the new parameter finder had the same walk
over a body's symbols written twice, and the second copy is the sort
that quietly stops matching the first.
